Use proper integer type

+++ nestkernel/sp_manager.cpp
@@ -717,7 +717,7 @@ SPManager::disable_structural_plasticity()
 GrowthCurve*
 SPManager::new_growth_curve( std::string name )
 {
-  const int nc_id = growthcurvedict_.get< long >( name );
+  const long nc_id = growthcurvedict_.get< long >( name );
   return growthcurve_factories_.at( nc_id )->create();
 }
